requirements.txt
# UI
# downgrade
PyQt6>=6.4.2

# LLM
google-genai

# Wake Word
openwakeword>=0.6.0

# STT
faster-whisper

# TTS 및 오디오 재생
edge-tts>=7.2.8
pygame>=2.6.0

# Utility
requests>=2.33.1
python-dotenv>=1.2.2

# bluetooth
pyserial

# Google API
google-api-python-client
google-auth-httplib2
google-auth-oauthlib

# conda
# conda install -c conda-forge alsa-plugins
# conda install -c conda-forge pyaudio portaudio